requirements.txt
hydra-core==1.2.0
numpy==1.23.1
omegaconf==2.2.2
pandas==1.4.3
pydantic==1.8.2
python-levenshtein==0.12.2
torch==1.12.0
scipy==1.8.1
spacy==3.3.1
thefuzz==0.19.0
tqdm==4.64.0
transformers==4.18.0